refactor: Repository에서 cartId 대신 productId를 사용하도록 리팩터링

+++ app/src/main/java/woowacourse/shopping/data/dto/request/OrderRequest.kt
@@ -1,5 +1,5 @@
 package woowacourse.shopping.data.dto.request
 
 data class OrderRequest(
-    val cartItemIds: List<Int>,
+    val cartItemIds: List<Long>,
 )
